--story=1006695 --user=龙儒锋 自动发布-应用配置-数据权限 逻辑修改 https://www.tapd.cn/54247054/s/1245407

+++ src/main/java/codedriver/module/deploy/api/appconfig/env/DeleteDeployAppConfigEnvDbConfigApi.java
@@ -3,6 +3,9 @@ package codedriver.module.deploy.api.appconfig.env;
 import codedriver.framework.auth.core.AuthAction;
 import codedriver.framework.common.constvalue.ApiParamType;
 import codedriver.framework.deploy.auth.DEPLOY_BASE;
+import codedriver.framework.deploy.constvalue.DeployAppConfigAction;
+import codedriver.framework.deploy.dto.app.DeployAppConfigEnvDBConfigVo;
+import codedriver.framework.deploy.exception.DeployAppConfigDbConfigNotFoundException;
 import codedriver.framework.restful.annotation.Description;
 import codedriver.framework.restful.annotation.Input;
 import codedriver.framework.restful.annotation.OperationType;
@@ -10,6 +13,7 @@ import codedriver.framework.restful.annotation.Param;
 import codedriver.framework.restful.constvalue.OperationTypeEnum;
 import codedriver.framework.restful.core.privateapi.PrivateApiComponentBase;
 import codedriver.module.deploy.dao.mapper.DeployAppConfigMapper;
+import codedriver.module.deploy.service.DeployAppAuthorityService;
 import com.alibaba.fastjson.JSONObject;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -28,6 +32,9 @@ public class DeleteDeployAppConfigEnvDbConfigApi extends PrivateApiComponentBase
 
     @Resource
     DeployAppConfigMapper deployAppConfigMapper;
+
+    @Resource
+    DeployAppAuthorityService deployAppAuthorityService;
 
     @Override
     public String getName() {
@@ -51,6 +58,15 @@ public class DeleteDeployAppConfigEnvDbConfigApi extends PrivateApiComponentBase
     @Override
     public Object myDoService(JSONObject paramObj) throws Exception {
         Long id = paramObj.getLong("id");
+        DeployAppConfigEnvDBConfigVo dbConfigVo = deployAppConfigMapper.getAppConfigEnvDBConfigById(id);
+        if (dbConfigVo == null) {
+            throw new DeployAppConfigDbConfigNotFoundException(id);
+        }
+
+        //校验环境权限、编辑配置的操作权限
+        deployAppAuthorityService.checkEnvAuth(dbConfigVo.getAppSystemId(), dbConfigVo.getEnvId());
+        deployAppAuthorityService.checkOperationAuth(dbConfigVo.getAppSystemId(), DeployAppConfigAction.EDIT);
+
 
         deployAppConfigMapper.deleteAppConfigDBConfigById(id);
         deployAppConfigMapper.deleteAppConfigDBConfigAccountByDBConfigId(id);
